mediaverse_Backend/requirements.txt
django
djangorestframework
psycopg2-binary
pillow
drf-spectacular
django-cors-headers
gunicorn==20.1.0